requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
graphene==3.4
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
packaging==24.1
promise==2.3
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
